Refactor metal commandbuffer
