+++ candle-metal-kernels/src/metal/command_buffer.rs
@@ -1,0 +1,53 @@
+use crate::{BlitCommandEncoder, ComputeCommandEncoder};
+use objc2::{rc::Retained, runtime::ProtocolObject};
+use objc2_foundation::NSString;
+use objc2_metal::{MTLCommandBuffer, MTLCommandBufferStatus};
+
+#[derive(Clone, Debug)]
+pub struct CommandBuffer {
+    raw: Retained<ProtocolObject<dyn MTLCommandBuffer>>,
+}
+
+impl CommandBuffer {
+    pub fn new(raw: Retained<ProtocolObject<dyn MTLCommandBuffer>>) -> Self {
+        Self { raw }
+    }
+
+    fn as_ref(&self) -> &ProtocolObject<dyn MTLCommandBuffer> {
+        &*self.raw
+    }
+
+    pub fn compute_command_encoder(&self) -> ComputeCommandEncoder {
+        self.as_ref()
+            .computeCommandEncoder()
+            .map(ComputeCommandEncoder::new)
+            .unwrap()
+    }
+
+    pub fn blit_command_encoder(&self) -> BlitCommandEncoder {
+        self.as_ref()
+            .blitCommandEncoder()
+            .map(BlitCommandEncoder::new)
+            .unwrap()
+    }
+
+    pub fn commit(&self) {
+        self.raw.commit()
+    }
+
+    pub fn enqueue(&self) {
+        self.raw.enqueue()
+    }
+
+    pub fn set_label(&self, label: &str) {
+        self.as_ref().setLabel(Some(&NSString::from_str(&label)))
+    }
+
+    pub fn status(&self) -> MTLCommandBufferStatus {
+        self.raw.status()
+    }
+
+    pub fn wait_until_completed(&self) {
+        unsafe { self.raw.waitUntilCompleted() }
+    }
+}
