+++ candle-metal-kernels/src/metal/mod.rs
@@ -1,9 +1,11 @@
 pub mod buffer;
+pub mod command_buffer;
 pub mod device;
 pub mod library;
 pub mod metal_utils;
 
 pub use buffer::*;
+pub use command_buffer::*;
 pub use device::*;
 pub use library::*;
 pub use metal_utils::*;
